Show the name of other users

+++ lib/web_rtc/signaling/hang_up.dart
@@ -20,45 +20,30 @@ extension HangUp on Signaling {
 
       List<Connection> connections = await room.connections();
 
-      connections.forEach((connection) async {
+      connections.asMap().forEach((index, connection) async {
         if (connection.source_user_id == user_id ||
             connection.destination_user_id == user_id) {
           DocumentReference connection_ref =
               room_ref.collection('connections').doc(connection.id);
 
-          _delete_connection_candidates(
+          await _delete_connection_candidates(
             connection_ref: connection_ref,
           );
           await connection_ref.delete();
+
+          if (room.host_id == user_id && index == connections.length - 1) {
+            await room_ref.delete();
+          }
         }
       });
 
-      if (room.host_id == user_id) {
-        await room_ref
-            .collection('connections')
-            .get()
-            .then((connections_query_snap) async {
-          connections_query_snap.docs
-              .asMap()
-              .forEach((index, connection_snap) async {
-            _delete_connection_candidates(
-              connection_ref: connection_snap.reference,
-            );
-            await connection_snap.reference.delete();
-
-            if (index == connections_query_snap.docs.length - 1) {
-              await room_ref.delete();
-            }
-          });
-        });
-      }
       remote_streams.forEach((remote_stream) {
         remote_stream.dispose();
       });
     }
   }
 
-  _delete_connection_candidates({
+  Future _delete_connection_candidates({
     required DocumentReference connection_ref,
   }) async {
     await connection_ref
